Moved database specification to separate app
